requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
lark==1.1.8
google-generativeai==0.3.2
python-decouple==3.8
python-multipart==0.0.6
sympy==1.12
langgraph>=0.2.0
langchain-core>=0.2.27
requests>=2.31.0